Move input interfaces to their own file
