+++ src/inputs.ts
@@ -1,0 +1,11 @@
+export interface UserInput {
+  name: string;
+  email: string;
+  password: string;
+  birthDate: string;
+}
+
+export interface LoginInput {
+  email: string;
+  password: string;
+}
