+++ src/schema.ts
@@ -32,15 +32,3 @@ export const typeDefs = `#graphql
     login(credentials: LoginInput): LoginInfo
   }
 `;
-
-export interface UserInput {
-  name: string;
-  email: string;
-  password: string;
-  birthDate: string;
-}
-
-export interface LoginInput {
-  email: string;
-  password: string;
-}
